Work around the diamond problem on Java 8
The AvlTree used to implemented both SortedSet and List. The diamond
problem arises in the class hierarchy when compiling / running on Java 8
because both interface now define the spliterator() method.

To work around the problem, the ADT now implements neither of the
interfaces and provides asList() and asSet() methods instead.