Harmonize visitor API
The TreeUtil's traversal methods did specify Visitor<TreeNode<T>>
instead of the less verbose Visitor<T>. That was not the initial design
idea, because for cases where access to the wrapping TreeNode is needed,
the methods that accept TreeNodeVisitor and TreeNodePrePostVisitor
should be used. The Visitor and PrePostVisitor deliberately hide the
TreeNode from the implementor to make implementing visitors easier in
cases where no access to the tree nodes is required.